Return default preferences when none are set
When the addon is activated through the Blender GUI and saved using the
"Save User Settings" it saves user preferences. When Blender is run from
the command line using the `--factory-startup` and the addon enabled
using `addon_utils.enable("animation_nodes")` there are no saved user
preferences.

The addon assumes that those preferences will be available and when
they're missing it causes errors.

This change wraps up the preferences that are checked in the course of
rendering into a `DefaultPreferences` object that it returns when no
preferences are available.